cat-app/py-backend-k8s/requirements.txt
flask==2.0.1
sqlalchemy==1.4.22
alembic==1.7.3
python-dotenv==0.19.0
requests==2.26.0
snowflake-sqlalchemy==1.2.4
six==1.16.0
flask_cors==3.0.10